Extract time transform into lib
